requirements.txt
Django==1.10.3
psycopg2==2.6.2
django-map-widgets==0.1.7
django-autoslug==1.9.3
